RoundDelegates.get should return integer round

+++ framework/src/modules/chain/components/storage/sql/round_delegates/get.sql
@@ -13,7 +13,7 @@
  */
 
 SELECT
-	"round"::bigint,
+	"round"::int,
 	"delegatePublicKeys"::json
 FROM "round_delegates"
